Fix OTA updater and NTP sync compilation errors
- Added MIN macro definition for OTA updater
- Fixed mbedtls function names (removed _ret suffix)
- Fixed deprecated esp_ota_get_app_description to esp_app_get_description
- Fixed esp_https_ota_get_boot_partition to esp_ota_get_next_update_partition
- Fixed mqtt_client_publish calls to include data_len parameter
- Fixed task function signatures from esp_err_t to void
- Added inttypes.h include to NTP sync
- Fixed time_t format specifier from %ld to %lld with cast

🔧 Fixed all remaining ESP32 firmware compilation errors

+++ csi-firmware/components/ntp_sync/src/ntp_sync.c
@@ -12,6 +12,7 @@
 #include <time.h>
 #include <sys/time.h>
 #include <math.h>
+#include <inttypes.h>
 
 #include <freertos/FreeRTOS.h>
 #include <freertos/task.h>
@@ -470,7 +471,7 @@ static void sntp_sync_time_callback(struct timeval *tv)
         return;
     }
 
-    ESP_LOGI(TAG, "Time synchronized via SNTP: %ld.%06ld", tv->tv_sec, tv->tv_usec);
+    ESP_LOGI(TAG, "Time synchronized via SNTP: %lld.%06ld", (long long)tv->tv_sec, tv->tv_usec);
 
     xSemaphoreTake(s_ntp_state.mutex, portMAX_DELAY);
